Textareas upload/download
Delete openFile function - this is in main.js now.
Delete doDL function - this downloading without filename, and buttons available always.
Add commented Guide tab, because this is exists, but empty. This tab not found in the code.

Add buttons to textareas for downloading is as text.
Now, when readonly textareas have text - there is possible to download it by click button.